+++ R/globals.R
@@ -1,0 +1,2 @@
+# adding tidyselect where
+utils::globalVariables("where")