+++ NEWS.md
@@ -1,5 +1,6 @@
 # RClimacell (development version)
 
+* added new function `climacell_core()` that retrieves all core layer data from Timeline Interface
 * added new function `climacell_celestial()` that retrieves the sunrise and sunset times along with the moon phase
 * removed the need to use the package {parsedate} and removed {parsedate} from dependency list
 * input dates no longer have to be in ISO8601 format (recommended to use the {lubridate} package or `Sys.time()` when entering date-time stamps)
@@ -8,7 +9,7 @@
 
 # RClimacell 0.1.3
 
-* fixed a bug where the user user provided API key was not used
+* fixed a bug where the user provided API key was not used
 
 # RClimacell 0.1.2